ref(trimTextCenter): Move trimTextCenter() into sentry/utils/string/*
