+++ static/app/utils/profiling/gl/utils.ts
@@ -19,6 +19,7 @@ import type {
 } from 'sentry/utils/profiling/renderers/UIFramesRenderer';
 import type {SpanChartNode} from 'sentry/utils/profiling/spanChart';
 import {Rect} from 'sentry/utils/profiling/speedscope';
+import type {TrimTextCenter} from 'sentry/utils/string/trimTextCenter';
 
 export function initializeFlamegraphRenderer(
   renderers: FlamegraphRendererConstructor[],
@@ -314,7 +315,6 @@ export function safeGetContext(
   return ctx;
 }
 
-export const ELLIPSIS = '\u2026';
 export function measureText(string: string, ctx?: CanvasRenderingContext2D): Rect {
   if (!string) {
     return Rect.Empty();
@@ -472,13 +472,6 @@ export function formatColorForFrame(
   }
 
   return `rgba(${color.map(n => n * 255).join(',')}, 1.0)`;
-}
-
-export interface TrimTextCenter {
-  end: number;
-  length: number;
-  start: number;
-  text: string;
 }
 
 export function hexToColorChannels(color: string, alpha: number): ColorChannels {
