+++ static/app/utils/profiling/renderers/flamegraphTextRenderer.tsx
@@ -7,7 +7,6 @@ import type {FlamegraphFrame} from 'sentry/utils/profiling/flamegraphFrame';
 import {getFlamegraphFrameSearchId} from 'sentry/utils/profiling/flamegraphFrame';
 import {
   computeHighlightedBounds,
-  ELLIPSIS,
   getContext,
   lowerBound,
   resizeCanvasToDisplaySize,
@@ -15,7 +14,9 @@ import {
 } from 'sentry/utils/profiling/gl/utils';
 import {TextRenderer} from 'sentry/utils/profiling/renderers/textRenderer';
 import type {Rect} from 'sentry/utils/profiling/speedscope';
-import {findRangeBinarySearch, trimTextCenter} from 'sentry/utils/profiling/speedscope';
+import {findRangeBinarySearch} from 'sentry/utils/profiling/speedscope';
+import {trimTextCenter} from 'sentry/utils/string/trimTextCenter';
+import {ELLIPSIS} from 'sentry/utils/string/unicode';
 
 class FlamegraphTextRenderer extends TextRenderer {
   flamegraph: Flamegraph;
