+++ static/app/utils/profiling/renderers/spansTextRenderer.tsx
@@ -4,7 +4,6 @@ import type {FlamegraphSearch} from 'sentry/utils/profiling/flamegraph/flamegrap
 import type {FlamegraphTheme} from 'sentry/utils/profiling/flamegraph/flamegraphTheme';
 import {
   computeHighlightedBounds,
-  ELLIPSIS,
   getContext,
   lowerBound,
   resizeCanvasToDisplaySize,
@@ -13,7 +12,9 @@ import {
 import {TextRenderer} from 'sentry/utils/profiling/renderers/textRenderer';
 import type {SpanChart, SpanChartNode} from 'sentry/utils/profiling/spanChart';
 import type {Rect} from 'sentry/utils/profiling/speedscope';
-import {findRangeBinarySearch, trimTextCenter} from 'sentry/utils/profiling/speedscope';
+import {findRangeBinarySearch} from 'sentry/utils/profiling/speedscope';
+import {trimTextCenter} from 'sentry/utils/string/trimTextCenter';
+import {ELLIPSIS} from 'sentry/utils/string/unicode';
 
 class SpansTextRenderer extends TextRenderer {
   spanChart: SpanChart;
