+++ static/app/utils/string/ellipsize.tsx
@@ -1,3 +1,5 @@
+import {ELLIPSIS} from 'sentry/utils/string/unicode';
+
 /**
  * Truncates a string to a maximum length and adds an ellipsis character (…) if truncated. Similar to the Python `truncatechars`. Does not trim the string, please trim before calling this function. Does right-trim the string before output to avoid having whitespace followed by an ellipsis, unless the string is only whitespace.
  *
@@ -22,5 +24,3 @@ export function ellipsize(str: string, length: number): string {
   // Otherwise, trim normally
   return `${str.slice(0, length).trimEnd()}${ELLIPSIS}`;
 }
-
-const ELLIPSIS = `\u2026`;
