+++ static/app/utils/string/unicode.ts
@@ -1,0 +1,1 @@
+export const ELLIPSIS = '\u2026';
